Fixing an error : when updating or deleting service images when the image is not found (user deleted, fs error, ...)
